+++ packages/environment-check/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/environment-check
 
+## 0.17.82
+
+### Patch Changes
+
+-   Updated dependencies [c8c292c]
+    -   @sap-ux/project-access@1.29.9
+    -   @sap-ux/axios-extension@1.18.6
+
 ## 0.17.81
 
 ### Patch Changes