+++ packages/fe-fpm-writer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/fe-fpm-writer
 
+## 0.33.7
+
+### Patch Changes
+
+-   Updated dependencies [c8c292c]
+    -   @sap-ux/project-access@1.29.9
+    -   @sap-ux/fiori-annotation-api@0.4.10
+
+## 0.33.6
+
+### Patch Changes
+
+-   0c37c3d: Added support for 'showClearButton', 'showMessages', and 'liveMode' properties in FilterBar building block generation.
+
 ## 0.33.5
 
 ### Patch Changes