+++ packages/fiori-elements-writer/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @sap-ux/fiori-elements-writer
 
+## 2.1.20
+
+### Patch Changes
+
+-   @sap-ux/annotation-generator@0.3.10
+-   @sap-ux/cap-config-writer@0.9.12
+-   @sap-ux/fe-fpm-writer@0.33.7
+-   @sap-ux/fiori-generator-shared@0.9.1
+-   @sap-ux/odata-service-writer@0.26.4
+-   @sap-ux/ui5-application-writer@1.2.10
+-   @sap-ux/ui5-test-writer@0.5.1
+
+## 2.1.19
+
+### Patch Changes
+
+-   Updated dependencies [f4867e5]
+    -   @sap-ux/ui5-test-writer@0.5.1
+
+## 2.1.18
+
+### Patch Changes
+
+-   Updated dependencies [fffc3a7]
+    -   @sap-ux/fiori-generator-shared@0.9.0
+    -   @sap-ux/ui5-test-writer@0.5.0
+    -   @sap-ux/cap-config-writer@0.9.11
+
+## 2.1.17
+
+### Patch Changes
+
+-   Updated dependencies [0c37c3d]
+    -   @sap-ux/fe-fpm-writer@0.33.6
+
 ## 2.1.16
 
 ### Patch Changes