+++ packages/inquirer-common/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/inquirer-common
 
+## 0.6.20
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.9.1
+-   @sap-ux/telemetry@0.5.63
+
+## 0.6.19
+
+### Patch Changes
+
+-   Updated dependencies [fffc3a7]
+    -   @sap-ux/fiori-generator-shared@0.9.0
+
 ## 0.6.18
 
 ### Patch Changes