+++ packages/ui5-test-writer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/ui5-test-writer
 
+## 0.5.1
+
+### Patch Changes
+
+-   f4867e5: Fix: Path Compatibility for Windows in FF Test Template Generation
+
+## 0.5.0
+
+### Minor Changes
+
+-   fffc3a7: Add Freestyle OPA templates to ui5-test-writer
+
 ## 0.4.1
 
 ### Patch Changes